requirements.txt
# Readning Project Requirements
# Python 3.11.x required (3.12 not yet supported)

# Core Framework
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic-settings>=2.10.1

# PyTorch and Audio Processing
# Note: These versions are specifically chosen for compatibility
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.41.2
numpy==1.26.3  # Compatible with torch 2.1.0

# Audio Processing
pydub>=0.25.1

# Text Processing & NLP
nltk>=3.9.1
beautifulsoup4>=4.12.3
ebooklib>=0.18
PyMuPDF>=1.24.0

# LLM Integration
ollama>=0.3.0

# File Handling
aiofiles>=24.1.0
python-multipart>=0.0.20

# Firebase Integration (optional)
firebase-admin>=6.5.0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.0

# Database (MySQL)
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0  # pymysql에서 SSL 지원용

# Audiocraft - Install separately from GitHub
# pip install git+https://github.com/facebookresearch/audiocraft.git

# For CUDA support (optional):
# CPU only:
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cpu
# 
# CUDA 11.8:
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu118
# 
# CUDA 12.1:
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121